requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-checkpoint-sqlite
langgraph-prebuilt
langchain-groq
langchain-huggingface
sentence-transformers
streamlit
faiss-cpu
pypdf
python-dotenv
requests
aiosqlite
duckduckgo-search
langchain-mcp-adapters
arxiv
ragas
pandas
datasets
pydantic
langsmith
